requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
Django==1.10.6
django-crispy-forms==1.6.1
django-extensions==1.7.8
django-filter==1.0.2
django-livereload-server==0.2.3
django-registration-redux==1.5
djangorestframework==3.6.2
google-api-python-client==1.6.2
httplib2==0.10.3
Markdown==2.6.8
oauth2client==4.0.0
packaging==16.8
pyasn1==0.2.3
pyasn1-modules==0.0.8
pydot==1.2.3
pyparsing==2.2.0
requests==2.13.0
rsa==3.4.2
six==1.10.0
tornado==4.4.3
uritemplate==3.0.0
